Move Some futures to utils module.
